trace-service/src/test/java/com/trace/service/UserIdConvertTest.java
package com.trace.service;

import org.junit.Assert;
import org.junit.Test;
import org.junit.runner.RunWith;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.boot.test.context.SpringBootTest;
import org.springframework.test.context.ContextConfiguration;
import org.springframework.test.context.junit4.SpringJUnit4ClassRunner;
import org.springframework.test.context.junit4.SpringRunner;

import static org.junit.Assert.*;

/**
 * @author xzp
 * Created on 2021/2/7
 */
@SpringBootTest
@RunWith(SpringJUnit4ClassRunner.class)
@ContextConfiguration(classes = SpringConfig.class)
public class UserIdConvertTest {
    @Autowired
    UserIdConvert convert;

    @Test
    public void doTest() {
        Integer res = convert.openIdConvert("123");
        Assert.assertNull(res);
    }
}